Add cleanup for gce

+++ publiccloud/azure/cleanup_images.pl
@@ -3,11 +3,18 @@ use Mojo::JSON qw(decode_json encode_json);
 use Date::Parse;
 use Mojo::File 'path';
 use Data::Dumper;
+use Sort::Versions;
+use Term::ANSIColor;
 
 my $rgroup = 'openqa-upload';
 my $storage_acc = 'openqa';
 my $storage_container = 'sle-images';
 my $keep_img = 1;
+my $dry_run = 0;
+
+
+sub color_delete { return color('bold red') . shift . color('reset'); }
+sub color_keep { return color('bold green') . shift . color('reset'); }
 
 sub run {
     my ($cmd) = @_;
@@ -24,13 +31,18 @@ sub annotate_images
     my $ret = {};
     for my $i (@{$data}){
         #            'name' => 'SLES12-SP5-Azure.x86_64-0.1.0-BYOS-Build1.14.vhd',
+        #           DELETE : SLES12-SP5-Azure.x86_64-0.9.0-SAP-BYOS-Build1.26.vhd
+        #
         my $name = $i->{name};
-        my ($prefix, $build) = $name =~ /^(SLES.+)-Build(\d+\.\d+)\.vhd$/;
+        my ($prefix, $kiwi, $prefix2, $build) = $name =~ /^(SLES.+)\.x86_64-(\d+\.\d+\.\d+)(.*)-Build(\d+\.\d+)\.vhd$/;
+        die("Failed to match $name") unless($kiwi);
+        $build = $kiwi . '-' . $build;
+        $prefix .= $prefix2 if ($prefix2);
         $img_by_group->{$prefix} //= [];
         push(@{$img_by_group->{$prefix}}, { name => $name, build => $build,  creationTime => str2time($i->{properties}->{creationTime}), delete_me => 0});
     }
     for my $key (keys %{$img_by_group}){
-        my @sorted_data = sort {$a->{build} cmp $b->{build}} @{$img_by_group->{$key}};
+        my @sorted_data = sort {versioncmp($a->{build},$b->{build})} @{$img_by_group->{$key}};
         for(my $i = 0; $i < @sorted_data; $i++){
 
             # Mark all older images to be deleted
@@ -50,20 +62,24 @@ my $images = annotate_images($data);
 
 while (my ($name, $img) = each %{$images}) {
     next if ($img->{delete_me} != 1);
-    print ("DELETE : " . $name . $/);
-    run("az storage blob delete --account-name $storage_acc --container-name $storage_container --name " . $name);
+    print (color_delete("DELETE " ) . $name . $/);
+    run("az storage blob delete --account-name $storage_acc --container-name $storage_container --name " . $name) unless $dry_run;
 }
 while (my ($name, $img) = each %{$images}) {
     next if ($img->{delete_me} != 0);
-    print ("KEEP : " . $name . $/);
+    print (color_keep("KEEP   ") . $name . $/);
 }
 
 $data = decode_json(run("az resource list --resource-group $rgroup"));
 for my $d (@{$data}){
-    next if (!exists($images->{$d->{name}}) || $images->{$d->{name}}->{delete_me} != 1);
+    if (!exists($images->{$d->{name}}) || $images->{$d->{name}}->{delete_me} != 1) {
+        next;
+    }
     if ($d->{'type'} eq 'Microsoft.Compute/images'){
-        run("az image delete --resource-group '$rgroup' --name '". $d->{name} . "'");
+        print color_delete("DELETE ") . " image $d->{name} $/";
+        run("az image delete --resource-group '$rgroup' --name '". $d->{name} . "'") unless $dry_run;
     } elsif ($d->{type} eq 'Microsoft.Compute/disks' ) {
-        run("az disk delete --resource-group '$rgroup' --name '". $d->{name} . "' -y");
+        print color_delete("DELETE ") . " disk $d->{name} $/";
+        run("az disk delete --resource-group '$rgroup' --name '". $d->{name} . "' -y") unless $dry_run;
     }
 }
